backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
paddlepaddle==2.6.2
paddleocr==2.9.1
pdf2image==1.17.0
Pillow==10.4.0
python-multipart==0.0.12
aiofiles==24.1.0
openpyxl==3.1.5
passlib[bcrypt]
bcrypt
pyjwt
email-validator
# ── Security additions ─────────────────────────────
python-dotenv>=1.0.0
slowapi>=0.1.9
